+++ SLK.Domain/Core/OrderItem.cs
@@ -4,13 +4,13 @@
     {
         protected OrderItem() { }
 
-        public long ID { get; protected set; }
+        public int ID { get; protected set; }
 
-        public long OrderID { get; protected set; }
+        public int OrderID { get; protected set; }
 
         public virtual Order Order { get; protected set; }
 
-        public long ProductInShopID { get; protected set; }
+        public int ProductInShopID { get; protected set; }
         
         public virtual ProductInShop ProductInShop { get; protected set; }
