+++ SLK.Domain/Core/ProductRate.cs
@@ -6,13 +6,13 @@ namespace SLK.Domain.Core
     {
         protected ProductRate() { }
 
-        public long ID { get; protected set; }
+        public int ID { get; protected set; }
 
-        public long UserID { get; protected set; }
+        public int UserID { get; protected set; }
 
         public virtual User User { get; protected set; }
 
-        public long ProductInShopID { get; protected set; }
+        public int ProductInShopID { get; protected set; }
 
         public virtual ProductInShop ProductInShop { get; protected set; }
