+++ SLK.Domain/Core/UserAddressSearch.cs
@@ -7,9 +7,9 @@ namespace SLK.Domain.Core
     {
         protected UserAddressSearch() { }
 
-        public long ID { get; protected set; }
+        public int ID { get; protected set; }
 
-        public long UserID { get; protected set; }
+        public int UserID { get; protected set; }
 
         public virtual User User { get; protected set; }
 
@@ -17,7 +17,7 @@ namespace SLK.Domain.Core
 
         public GeoCoordinate Coordinate { get; protected set; }
 
-        public long ShopID { get; protected set; }
+        public int ShopID { get; protected set; }
 
         public virtual Shop Shop { get; protected set; }
 
@@ -29,7 +29,7 @@ namespace SLK.Domain.Core
 
         public string WrotenAddress { get; protected set; }
 
-        public long ShopTypeID { get; protected set; }
+        public int ShopTypeID { get; protected set; }
 
         public virtual ShopType ShopType { get; protected set; }
     }
